Listener toggle command. Closes #7

+++ index.js
@@ -10,8 +10,8 @@ const path = require('node:path');
 const { token, MONGO_URI } = require('./config.json');
 const { Client, GatewayIntentBits, Collection } = require('discord.js');
 const mongoose = require('mongoose');
-
-const testSchema = require('./test-schema');
+const database = require('./data/schema/replace-toggle');
+const Twitter = require('./twitter/twitter-api')
 
 // Create client instance
 const client = new Client({
@@ -42,11 +42,6 @@ client.once('ready', async () => {
 			keepAlive: true
 		}
 	)
-
-	await new testSchema({
-		message: "hello world",
-	}).save()
-	
 	console.log('S.N.E.K is Ready!');
 });
 
@@ -71,13 +66,10 @@ client.on('interactionCreate', async interaction => {
 
 });
 
-// Log in the client
-client.login(token);
-
-//Read messages and replace any twitter links containing videos with a vxtwitter link
-/* 
-client.on('messageCreate', message => {
-
+//on messages being created
+client.on('messageCreate', async message => {
+	
+	//if message contains twitter link
 	if (message.content.includes('https://twitter.com')) {
 		// console.log('Twitter link detected'); // debug use
 
@@ -94,19 +86,26 @@ client.on('messageCreate', message => {
 		const checkTweet = async () => {
 			try {
 				// Make request
-				const response = await getRequest(tweetId);
+				const response = await Twitter.getRequest(tweetId);
 				if (response.includes.media[0].type === 'video') {
-					// console.log(response.includes.media[0].type); // debug
-					// console.log('tweet contains video, replacing link'); // debug
 
-					// delete the original message
-					message.delete();
+					const replaceDB = await database.findOne({id: message.guildId});
+					//if Guild is in database then replace tweet
+					if (!replaceDB.toggled === false)
+					{
+						// console.log(response.includes.media[0].type); // debug
+						// console.log('tweet contains video, replacing link'); // debug
 
-					// send message to denote which user sent the link
-					message.channel.send(`${username} posted:`, { 'allowedMentions': { 'users' : [] } });
+						// delete the original message
+						message.delete();
 
-					// send the new vxtwitter message
-					message.channel.send(`https://vxtwitter.com/${twitterUser}/status/${tweetId}`);
+						// send message to denote which user sent the link
+						message.channel.send(`${username} posted:`, { 'allowedMentions': { 'users' : [] } });
+
+						// send the new vxtwitter message
+						message.channel.send(`https://vxtwitter.com/${twitterUser}/status/${tweetId}`);	
+					}
+					
 				}
 				else {
 					console.log('no video detected, letting discord embed twitter link');
@@ -122,4 +121,6 @@ client.on('messageCreate', message => {
 		checkTweet();
 	}
 });
- */
+
+// Log in the client
+client.login(token);
